Refactor datasets to make it easier to make more, create cache decorator
