+++ datasets/sba_city_county.py
@@ -1,0 +1,28 @@
+URL = 'http://api.sba.gov/geodata/city_county_links_for_state_of/{}.json'
+
+STATE_CODES = [
+    "AL", "AK", "AZ", "AR", "CA", "CO", "CT", "DC", "DE", "FL", "GA",
+    "HI", "ID", "IL", "IN", "IA", "KS", "KY", "LA", "ME", "MD",
+    "MA", "MI", "MN", "MS", "MO", "MT", "NE", "NV", "NH", "NJ",
+    "NM", "NY", "NC", "ND", "OH", "OK", "OR", "PA", "RI", "SC",
+    "SD", "TN", "TX", "UT", "VT", "VA", "WA", "WV", "WI", "WY"
+]
+
+def _grab_state_data(state_code):
+    logging.info('Looking up ---%s---', state_code)
+    response = requests.get(URL.format(state_code))
+    try:
+        response = response.json()
+        return {
+            city['name']: (city['fips_county_cd'], city['county_name'])
+            for city in response
+        }
+    except:
+        return {}
+
+@cacheable('county_lookup.json')
+def county_lookup():
+    return {
+        state_code: _grab_state_data(state_code)
+        for state_code in STATE_CODES
+    }
